+++ docs/changes.rst
@@ -8,6 +8,7 @@ Released on ? 2016.
 
 * Removed Google web translate machine translation.
 * Improved commit message when adding translation.
+* Fixed Google Translate API for Hebrew language.
 
 weblate 2.6
 -----------